Makes recursive-copy return a failure exit code on error.
When an error occurs during the `recursive-copy` script, it should propagate that error up to the calling shell so it can branch on it.  For example, in `recrusive-copy source/foo destination/foo && ./destination/foo/script.js` the second command should not execute if the first command fails.  Right now, the first command will fail but return a 0 exit code, resulting in the following command attempting to run (and possibly failing).